requirements.txt
flask
gunicorn
requests
paho-mqtt
yfinance
pandas
numpy
python-dotenv
